+++ Android/app/src/main/AndroidManifest.xml
@@ -9,6 +9,7 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
+    <uses-permission android:name="android.permission.CAMERA"/>
 
     <application
         android:name=".myApplication.MyApplication"
@@ -24,7 +25,7 @@
             android:name=".Sections.materialdesign.MaterialDesignActivity"
             android:exported="true" />
         <activity
-            android:name=".recoder.RecoderActivity"
+            android:name=".sections.recoder.RecoderActivity"
             android:exported="true" />
 
         <service
