+++ Android/app/src/main/res/layout/activity_recoder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".recoder.RecoderActivity">
+    tools:context=".sections.recoder.RecoderActivity">
 
     <TextureView
         android:id="@+id/textureView"
